refactor: consolidate waitlist logic into events/waitlist module
